moves html and json creation functions into new files
